bot/requirements.txt
# Telegram Bot
aiogram==3.7.0
python-dotenv==1.0.0

# PostgreSQL + SQLAlchemy
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0

# Дополнительно
python-multipart==0.0.9
httpx==0.27.0
